requirements.txt
boto3>=1.40.25
fastapi==0.116.1
langchain-core==0.3.75
langchain-aws==0.2.31
langchain-community==0.3.30
langchain-redis==0.2.4
langchain-qdrant==0.2.1
uvicorn==0.38.0
pyyaml==6.0.2